+++ hc/api/management/commands/prunenotifications.py
@@ -12,6 +12,7 @@ class Command(BaseCommand):
 
         q = Check.objects.filter(n_pings__gt=100)
         q = q.annotate(min_ping_date=Min("ping__created"))
+        q = q.exclude(min_ping_date=None)
         for check in q:
             qq = Notification.objects.filter(
                 owner_id=check.id, created__lt=check.min_ping_date