+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 
 ### Bug Fixes
 - Fix unwanted localization in badge SVG generation (#629)
+- Update email template to handle not yet uploaded ping bodies
 
 ## v2.0.1 - 2022-03-18