+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 - Implement the "started" progress spinner in the details pages
 - Add "hc_check_started" metric in the Prometheus metrics endpoint (#630)
 - Add a management command for submitting Signal rate limit challenges
+- Upgrade to django-compressor 4.0
 
 ### Bug Fixes
 - Fix unwanted localization in badge SVG generation (#629)